统一db 表字段及部分表名变更

Created-by: msprofDeveloper

Match-id-b42e4d5f899c3bcf272f2baf8c63712d72ff286f

+++ analysis/csrc/viewer/database/finals/llc_processor.h
@@ -24,7 +24,7 @@ public:
     // device_id, l3tid, timestamp, hitrate, throughput
     using OriDataFormat = std::vector<std::tuple<uint32_t, uint32_t, double, double, double>>;
     // deviceId, llcID, timestamp, hitRate, throughput, mode
-    using ProcessedDataFormat = std::vector<std::tuple<uint16_t, uint32_t, uint64_t, double, uint64_t, uint16_t>>;
+    using ProcessedDataFormat = std::vector<std::tuple<uint16_t, uint32_t, uint64_t, double, uint64_t, uint64_t>>;
 
     LLCProcessor() = default;
     LLCProcessor(const std::string &reportDBPath, const std::set<std::string> &profPaths);
@@ -37,7 +37,7 @@ protected:
 private:
     static OriDataFormat GetData(const std::string &dbPath, DBInfo &dbInfo);
     static ProcessedDataFormat FormatData(const OriDataFormat &oriData, const ThreadData &threadData,
-                                          const uint16_t mode);
+                                          const uint64_t mode);
     bool ProcessData(const std::string &fileDir);
 };
 } // Database
